refactor code and update documentation
