+++ examples/image_text_to_text/README.md
@@ -91,6 +91,7 @@ Popular model families include:
 - InternVL
 - Molmo
 - LLaVA
+- Kimi-K2.5
 
 ### Model-Specific Examples
 
@@ -107,6 +108,21 @@ Some models have specialized examples demonstrating advanced features:
 | **Granite** | [models/granite_vision/](models/granite_vision/) |
 | **InternVL** | [models/internvl/](models/internvl/) |
 | **Molmo** | [models/molmo/](models/molmo/) |
+| **Kimi-K2.5** | [../kimi_k2/export_kimi_k25_vision.py](../kimi_k2/export_kimi_k25_vision.py) |
+
+### Kimi-K2.5 Vision
+
+Kimi-K2.5 uses the `moonshotai/Kimi-K2.5` remote-code model with the `KimiK25ForConditionalGeneration` architecture. The example script loads either the full checkpoint or a smaller layer subset, wraps it with `QEFFAutoModelForImageTextToText`, and compiles the vision-language path with Kimi-specific image-grid specializations.
+
+```bash
+
+python ../kimi_k2/export_kimi_k25_vision.py \
+    --model-path "$HF_HUB_CACHE/models--moonshotai--Kimi-K2.5/snapshots/<snapshot-id>" \
+    --image-url "https://huggingface.co/moonshotai/Kimi-K2.5/resolve/main/figures/kimi-logo.png" \
+    --prompt "Describe this image."
+```
+
+For custom image sizes, pass matching compile specializations for the processed image grid: `num_patches`, `h`, `w`, and `num_image_tokens`. The defaults in `../kimi_k2/export_kimi_k25_vision.py` demonstrate the expected values for the sample image and keep the language-side image embedding shape bounded to the actual image token count.
 
 For reranker examples, see [../reranker/](../reranker/).
